вынесли логику в кастомный хук useGenderize, перевели полученные данные на ru, добавили проверку на пустое имя, вынесли интерфейс в отдельный файл
