+++ src/Types.ts
@@ -1,0 +1,6 @@
+interface GenderData {
+  name: string;
+  gender: string;
+}
+
+export default GenderData;
